Rename: 파일의 경로 수정
UserAccount 를 entity 패키지로 이동함

Related to: #5
