+++ src/main/java/pulleydoreurae/chwijunjindan/auth/domain/entity/UserAccount.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.chwijunjindan.auth.domain;
+package pulleydoreurae.chwijunjindan.auth.domain.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -10,6 +10,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import pulleydoreurae.chwijunjindan.auth.domain.UserRole;
 
 /**
  * 회원가입이나 로그인 시 사용자 정보를 담을 엔티티
