+++ src/main/java/pulleydoreurae/chwijunjindan/auth/service/KakaoLoginService.java
@@ -19,7 +19,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 import com.google.gson.Gson;
 
 import lombok.extern.slf4j.Slf4j;
-import pulleydoreurae.chwijunjindan.auth.domain.UserAccount;
+import pulleydoreurae.chwijunjindan.auth.domain.entity.UserAccount;
 import pulleydoreurae.chwijunjindan.auth.domain.UserRole;
 import pulleydoreurae.chwijunjindan.auth.domain.dto.KakaoLoginResponse;
 import pulleydoreurae.chwijunjindan.auth.domain.dto.KakaoUserDetailsResponse;
